Move useEditNameAndProduction to ProductionStep
